refactoring: entity packacge created inside model
