+++ src/main/java/org/itmo/eventapp/main/model/entity/PrivilegeType.java
@@ -1,4 +1,4 @@
-package org.itmo.eventapp.main.model;
+package org.itmo.eventapp.main.model.entity;
 
 public enum PrivilegeType {
     APPROVE_REGISTRATION_REQUEST,
